Resolve an issue with ABSPATH in plugins.
A common security practice for WordPress plugins is to include a check for ABSPATH at the top of files and exit early if it does not exist.

Here is the most common pattern:

```
if ( ! defined( 'ABSPATH' ) ) {
	exit; // Exit if accessed directly
}
```

That guard was causing strauss to exit before processing anything. This patch adds that constant back so the plugin code can be loaded without stopping strauss execution.